PhoneField styled-component migration
